Make hacheck look for host in headers
This makes it possible to override the host to check from two
headers:
 * X-Haproxy-Server-State which is sent by HAProxy and we already use to
override the check port
 * X-Nerve-Check-IP which I am going to configure Nerve to send

The default is to fallback to checking localhost as we do now. The
motivation for this change is that Kubernetes Pods get their own IP. So
we currently have to run hacheck in a sidecar for every service we run
since hacheck currently hardcodes checking localhost. With this change
we can configure Nerve to send this new header and use 1 system hacheck
for all the Pods on a host.