Add static build of haupdown
This is kinda nuts so apologies...

I'm trying to build a statically linked binary version of haupdown that
I can run anywhere. For example, in a kubernetes service container via a
volume mount to the binary on the host. This way kubernetes container
hooks can call haupdown without needing hacheck in the service container
or in a sidecar.

Any suggestions on a cleaner way to achieve that are welcomed, this was
just the first thing I tried that worked.